Refactor progress window to separate files
